Added improvements to Kibana tests
