+++ .vib/kibana/chart-values.yaml
@@ -1,0 +1,18 @@
+extraConfiguration:
+  csp.strict: false
+service:
+  ports: 
+    http: 80
+  type: LoadBalancer
+elasticsearch:
+  hosts:
+    - localhost
+  port: 9200
+sidecars: 
+ - name: elasticsearch
+   image: bitnami/elasticsearch:latest
+   imagePullPolicy: IfNotPresent
+   ports:
+     - name: http
+       containerPort: 9200
+
