+++ .vib/kibana/cypress/cypress/fixtures/clusters.json
@@ -1,0 +1,6 @@
+{
+  "newCluster": {
+    "name": "remoteCluster",
+    "cluter": "localhost:90"
+  }
+}
