+++ .vib/kibana/cypress/cypress/fixtures/dashboards.json
@@ -1,0 +1,6 @@
+{
+  "newDashboard": {
+    "title": "Test_dashboard",
+    "description": "Test_visualisation"
+  }
+}
