+++ .vib/kibana/cypress/cypress/fixtures/indexes.json
@@ -1,0 +1,5 @@
+{
+  "newIndex": {
+    "name": "test_index"
+  }
+}
