+++ .vib/kibana/cypress/cypress/fixtures/visualizations.json
@@ -1,0 +1,6 @@
+{
+  "newVisualization": {
+    "title": "Test_dashboard",
+    "description": "Test_visualisation"
+  }
+}
