+++ .vib/kibana/cypress/cypress/fixtures/workpads.json
@@ -1,0 +1,5 @@
+{
+  "newWorkpad": {
+    "name": "testWorkpad"
+  }
+}
